Move grade constants to `common`
